- DispatchResult::route() moved to Collector::route() - DispatchResult::uri() moved to Collector::uri() - README.md & CHANGELOG.md update
