+++ src/Router/Collector.php
@@ -6,6 +6,7 @@ namespace QuickRoute\Router;
 use FastRoute\DataGenerator\GroupCountBased;
 use FastRoute\RouteCollector as FastRouteCollector;
 use FastRoute\RouteParser\Std;
+use InvalidArgumentException;
 use QuickRoute\Route;
 
 class Collector
@@ -279,4 +280,57 @@ class Collector
     {
         return $this->isRegistered;
     }
+
+
+    /**
+     * Find route by name
+     *
+     * @param string $routeName
+     * @return array|null
+     */
+    public function route(string $routeName): ?array
+    {
+        $this->doCollectRoutes();
+        foreach ($this->collectedRoutes as $collectedRoute) {
+            if ($routeName == $collectedRoute['name']) {
+                return $collectedRoute;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Generate route http uri using route's name
+     *
+     * @param string $routeName
+     * @param array $routeParams an array of [key => value] of route parameters
+     * @return string|null
+     */
+    public function uri(string $routeName, array $routeParams = []): ?string
+    {
+        $foundRoute = $this->route($routeName);
+        if (!$foundRoute) return null;
+
+        $prefix = $foundRoute['prefix'] ?? null;
+        if (!$prefix) return null;
+
+        return $this->replaceParamWithValue($prefix, $routeParams);
+    }
+
+    protected function replaceParamWithValue(string $prefix, array $params): string
+    {
+        preg_match_all("@{([0-9a-zA-Z]+):?.*?\+?}@", $prefix, $matchedParams);
+
+        for ($i = 0; $i < count($matchedParams[0]); $i++) {
+            $paramValue = $params[$matchedParams[1][$i]] ?? null;
+            if (null == $paramValue) {
+                throw new InvalidArgumentException("Missing route parameter value for \"{$matchedParams[1][$i]}\"");
+            }
+            $prefix = str_replace($matchedParams[0][$i], $params[$matchedParams[1][$i]], $prefix);
+        }
+
+        return $prefix;
+    }
+
 }
