+++ src/Router/DispatchResult.php
@@ -14,17 +14,18 @@ class DispatchResult
      */
     private array $dispatchResult;
 
-    private array $collectedRoutes;
+    private Collector $collector;
 
     /**
      * DispatchResult constructor.
      *
      * @param string[] $dispatchResult
+     * @param Collector $collector
      */
-    public function __construct(array $dispatchResult, array $collectedRoutes)
+    public function __construct(array $dispatchResult, Collector $collector)
     {
         $this->dispatchResult = $dispatchResult;
-        $this->collectedRoutes = $collectedRoutes;
+        $this->collector = $collector;
     }
 
 
@@ -68,63 +69,13 @@ class DispatchResult
     }
 
     /**
-     * Find route by name
-     *
-     * @param string $routeName
-     * @return array|null
-     */
-    public function route(string $routeName): ?array
-    {
-        foreach ($this->collectedRoutes as $collectedRoute) {
-            if ($routeName == $collectedRoute['name']) {
-                return $collectedRoute;
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Generate route http uri using route's name
-     *
-     * @param string $routeName
-     * @param array $routeParams an array of [key => value] of route parameters
-     * @return string|null
-     */
-    public function uri(string $routeName, array $routeParams = []): ?string
-    {
-        $foundRoute = $this->route($routeName);
-        if (!$foundRoute) return null;
-
-        $prefix = $foundRoute['prefix'] ?? null;
-        if (!$prefix) return null;
-
-        return $this->replaceParamWithValue($prefix, $routeParams);
-    }
-
-    protected function replaceParamWithValue(string $prefix, array $params): string
-    {
-        preg_match_all("@{([0-9a-zA-Z]+):?.*?\+?}@", $prefix, $matchedParams);
-
-        for ($i = 0; $i < count($matchedParams[0]); $i++) {
-            $paramValue = $params[$matchedParams[1][$i]] ?? null;
-            if (null == $paramValue) {
-                throw new InvalidArgumentException("Missing route parameter value for \"{$matchedParams[1][$i]}\"");
-            }
-            $prefix = str_replace($matchedParams[0][$i], $params[$matchedParams[1][$i]], $prefix);
-        }
-
-        return $prefix;
-    }
-
-    /**
      * Get all collected routes
      *
-     * @return array
+     * @return Collector
      */
-    public function getCollectedRoutes(): array
+    public function getCollector(): Collector
     {
-        return $this->collectedRoutes;
+        return $this->collector;
     }
 
     /**
